♻️ (#350) 유틸과 컴포넌트 분리
