+++ src/features/webpush/components/StatusView.tsx
@@ -1,0 +1,38 @@
+import { cn } from '@/shared/lib/utils';
+
+interface StatusViewProps {
+  title: string;
+  message: React.ReactNode;
+  icon: React.ComponentType<{ className?: string }>;
+  blurClass: string;
+  bgClass: string;
+  textClass: string;
+  children?: React.ReactNode;
+}
+
+const StatusView = ({
+  title,
+  message,
+  icon: Icon,
+  blurClass,
+  bgClass,
+  textClass,
+  children,
+}: StatusViewProps) => (
+  <div className="flex flex-col h-screen justify-center items-center text-center p-6 space-y-6">
+    <div className="relative inline-block">
+      <div className={cn('absolute inset-0 rounded-full blur-xs', blurClass)} />
+      <div className={cn('relative p-4 rounded-full w-fit mx-auto', bgClass)}>
+        <Icon className={cn('w-10 h-10', textClass)} />
+      </div>
+    </div>
+
+    <div className="space-y-2">
+      <h2 className="title1-bold text-gray-900">{title}</h2>
+      <p className="subtitle2-regular text-gray-600 leading-relaxed">{message}</p>
+    </div>
+
+    {children}
+  </div>
+);
+export default StatusView;
