+++ src/features/webpush/utils/deviceUtil.ts
@@ -1,0 +1,7 @@
+export const getIsIOS = () => {
+  if (typeof window === 'undefined') return false;
+  return /iPhone|iPad|iPod/.test(navigator.userAgent);
+};
+export const getIsStandalone = () =>
+  window.matchMedia('(display-mode: standalone)').matches ||
+  (navigator as unknown as { standalone?: boolean }).standalone === true;
